X-loader 512 byte SW ECC support for Zoom1 and Zoom2
Patches add standard 512 byte ecc support in xloader and u-boot for s/w ecc.
This is compatible with what kernel has.

These are needed for flashing of android filesystem using fastboot so that
kernel recognizes the data flashed from fastboot.

512 byte ecc is supported only in omap3430labrador and omap3430zoom2.
This is with the board config variable CFG_SW_ECC_512

If CFG_SW_ECC_512 is 512, the new ecc algorithm is used, if not the
old algorithm is used.

Run tested on Zoom1
Compile tested on Zoom2 and omap3430sdp

Signed-off-by: Vimal Singh <Vimal.Singh@ti.com>
Signed-off-by: Tom Rix <Tom.Rix@windriver.com>
Acked-by: Nishanth Menon <nm@ti.com>