X-loader update for cpu revision change in u-boot.
These changes are necessary because the header files in u-boot changed.
The get_cpu_rev changes in u-boot were mirrored in x-loader.
The board specific get_cpu_rev was moved to the new cpu level sys_info.c file.
cpu_is_3410 was also common so it was moved as well.

Signed-off-by: Tom Rix <Tom.Rix@windriver.com>