OMAP4: clocks: Enable only required clks
X-loader untill now, was enabling all clks at bootup
to help all modules to be functional at the kernel, even
with drivers which do not handle clks well.
Now that we are moving towards all drivers being adapted
to request/release clks as expected, most of this code is
useless and hence removed.

Signed-off-by: Rajendra Nayak <rnayak@ti.com>